Fix(Booking): Return BookingId when create new one. Fix(Payment): Replace type price.

+++ Core/Fieldy.BookingYard.Application/Features/Payment/Commands/CreatePayment/CreatePaymentCommand.cs
@@ -5,6 +5,6 @@ namespace Fieldy.BookingYard.Application.Features.Payment.Commands.CreatePayment
 	public record CreatePaymentCommand: IRequest<string>
 	{
 		public Guid BookingId { get; set; }
-		public int RequiredAmount { get; set; }
+		public decimal RequiredAmount { get; set; }
 	}
 }
